# perf(model): remove attr iterations when creating a model instance