@makepanic 2019-10-16
rebased on current master